feat: added tracing with go openai client
