+++ examples/otel_anthropic/traceanthropic/traceanthropic.go
@@ -1,0 +1,325 @@
+// Package traceanthropic provides OpenTelemetry tracing for the
+// Anthropic API client using LangSmith-compatible spans.
+//
+// Usage:
+//
+//	// Configure your Anthropic client to use a traced HTTP client
+//	client := anthropic.NewClient(
+//		anthropic.WithAPIKey(apiKey),
+//		anthropic.WithHTTPClient(traceanthropic.Client()),
+//	)
+//
+//	// Your Anthropic API calls will now be automatically traced with LangSmith attrs
+//	// resp, err := client.Messages.Create(ctx, ...)
+package traceanthropic
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+	"strings"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/propagation"
+	"go.opentelemetry.io/otel/trace"
+)
+
+// Client returns a new http.Client configured with tracing middleware.
+// Equivalent to WrapClient(nil), which wraps the default transport.
+func Client() *http.Client {
+	return WrapClient(nil)
+}
+
+// WrapClient wraps an existing http.Client with tracing middleware.
+// If client is nil, a new client with the default transport is created.
+func WrapClient(client *http.Client) *http.Client {
+	if client == nil {
+		client = &http.Client{}
+	}
+	transport := client.Transport
+	if transport == nil {
+		transport = http.DefaultTransport
+	}
+	client.Transport = newRoundTripper(transport)
+	return client
+}
+
+type roundTripper struct {
+	base http.RoundTripper
+}
+
+func newRoundTripper(base http.RoundTripper) http.RoundTripper {
+	return &roundTripper{base: base}
+}
+
+// RoundTrip intercepts requests/responses to add OpenTelemetry tracing.
+func (rt *roundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
+	// Only trace Anthropic API requests
+	if !strings.Contains(req.URL.Host, "api.anthropic.com") {
+		return rt.base.RoundTrip(req)
+	}
+
+	ctx := req.Context()
+	tracer := otel.Tracer("github.com/anthropics/anthropic-sdk-go")
+
+	// Extract span context from request headers
+	ctx = otel.GetTextMapPropagator().Extract(ctx, propagation.HeaderCarrier(req.Header))
+
+	// Determine span name based on endpoint
+	spanName := getSpanName(req.URL.Path)
+
+	// Start span
+	ctx, span := tracer.Start(ctx, spanName,
+		trace.WithAttributes(
+			attribute.String("gen_ai.system", "anthropic"),
+			attribute.String("gen_ai.operation.name", getOperationName(req.URL.Path)),
+			attribute.String("http.method", req.Method),
+			attribute.String("http.url", req.URL.String()),
+		),
+	)
+	defer span.End()
+
+	// Read request body if present
+	var requestBody []byte
+	if req.Body != nil {
+		var err error
+		requestBody, err = io.ReadAll(req.Body)
+		if err != nil {
+			span.RecordError(err)
+			span.SetStatus(codes.Error, fmt.Sprintf("failed to read request body: %v", err))
+			return rt.base.RoundTrip(req)
+		}
+		req.Body = io.NopCloser(bytes.NewBuffer(requestBody))
+	}
+
+	// Extract request attributes from body
+	if len(requestBody) > 0 {
+		extractRequestAttributes(span, requestBody)
+	}
+
+	// Inject span context into request headers
+	otel.GetTextMapPropagator().Inject(ctx, propagation.HeaderCarrier(req.Header))
+
+	// Make the actual request
+	resp, err := rt.base.RoundTrip(req)
+	if err != nil {
+		span.RecordError(err)
+		span.SetStatus(codes.Error, err.Error())
+		return resp, err
+	}
+
+	// Read response body
+	var responseBody []byte
+	if resp.Body != nil {
+		var readErr error
+		responseBody, readErr = io.ReadAll(resp.Body)
+		if readErr != nil {
+			span.RecordError(readErr)
+			span.SetStatus(codes.Error, fmt.Sprintf("failed to read response body: %v", readErr))
+			// Continue with empty body rather than failing the request
+		} else {
+			resp.Body = io.NopCloser(bytes.NewBuffer(responseBody))
+		}
+	}
+
+	// Extract response attributes
+	if len(responseBody) > 0 {
+		extractResponseAttributes(span, responseBody)
+	}
+
+	// Set status based on HTTP status code
+	if resp.StatusCode >= 400 {
+		span.SetStatus(codes.Error, fmt.Sprintf("HTTP %d", resp.StatusCode))
+	} else {
+		span.SetStatus(codes.Ok, "")
+	}
+
+	return resp, err
+}
+
+// getSpanName returns an appropriate span name based on the API endpoint.
+func getSpanName(path string) string {
+	if strings.Contains(path, "/v1/messages") {
+		return "anthropic.messages"
+	}
+	return "anthropic.request"
+}
+
+// getOperationName returns the operation name for Gen AI semantic conventions.
+func getOperationName(path string) string {
+	if strings.Contains(path, "/v1/messages") {
+		return "chat"
+	}
+	return "request"
+}
+
+// extractRequestAttributes extracts attributes from Anthropic request body.
+func extractRequestAttributes(span trace.Span, body []byte) {
+	var req map[string]interface{}
+	if err := json.Unmarshal(body, &req); err != nil {
+		return
+	}
+
+	// Extract model
+	if model, ok := req["model"].(string); ok {
+		span.SetAttributes(attribute.String("gen_ai.request.model", model))
+	}
+
+	// Extract max_tokens
+	if maxTokens, ok := req["max_tokens"].(float64); ok {
+		span.SetAttributes(attribute.Int64("gen_ai.request.max_tokens", int64(maxTokens)))
+	}
+
+	// Extract temperature
+	if temp, ok := req["temperature"].(float64); ok {
+		span.SetAttributes(attribute.Float64("gen_ai.request.temperature", temp))
+	}
+
+	// Extract system message if present
+	var messages []interface{}
+	if sys, ok := req["system"]; ok {
+		messages = append(messages, map[string]interface{}{
+			"role":    "system",
+			"content": sys,
+		})
+	}
+
+	// Extract user messages
+	if msgs, ok := req["messages"].([]interface{}); ok {
+		messages = append(messages, msgs...)
+	}
+
+	// Build prompt from messages
+	if len(messages) > 0 {
+		prompt := buildPromptFromMessages(messages)
+		if prompt != "" {
+			span.SetAttributes(attribute.String("gen_ai.prompt", prompt))
+		}
+	}
+}
+
+// buildPromptFromMessages builds a prompt string from Anthropic messages.
+func buildPromptFromMessages(messages []interface{}) string {
+	var parts []string
+	for _, msg := range messages {
+		if msgMap, ok := msg.(map[string]interface{}); ok {
+			if role, _ := msgMap["role"].(string); role == "user" || role == "system" {
+				if content, ok := msgMap["content"].(string); ok {
+					parts = append(parts, content)
+				} else if contentArray, ok := msgMap["content"].([]interface{}); ok {
+					// Handle content blocks
+					for _, block := range contentArray {
+						if blockMap, ok := block.(map[string]interface{}); ok {
+							if blockType, _ := blockMap["type"].(string); blockType == "text" {
+								if text, _ := blockMap["text"].(string); text != "" {
+									parts = append(parts, text)
+								}
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+	return strings.Join(parts, "\n")
+}
+
+// extractResponseAttributes extracts attributes from Anthropic response body.
+func extractResponseAttributes(span trace.Span, body []byte) {
+	var resp map[string]interface{}
+	if err := json.Unmarshal(body, &resp); err != nil {
+		return
+	}
+
+	// Extract model
+	if model, ok := resp["model"].(string); ok {
+		span.SetAttributes(attribute.String("gen_ai.response.model", model))
+	}
+
+	// Extract usage
+	if usage, ok := resp["usage"].(map[string]interface{}); ok {
+		setUsageAttributes(span, usage)
+	}
+
+	// Extract stop_reason
+	if stopReason, ok := resp["stop_reason"].(string); ok && stopReason != "" {
+		span.SetAttributes(attribute.String("langsmith.metadata.stop_reason", stopReason))
+	}
+
+	// Extract completion from content
+	if content, ok := resp["content"].([]interface{}); ok && len(content) > 0 {
+		completion := extractCompletionFromContent(content)
+		if completion != "" {
+			span.SetAttributes(attribute.String("gen_ai.completion", completion))
+		}
+	}
+}
+
+// extractCompletionFromContent extracts text completion from Anthropic content blocks.
+func extractCompletionFromContent(content []interface{}) string {
+	var parts []string
+	for _, block := range content {
+		if blockMap, ok := block.(map[string]interface{}); ok {
+			if blockType, _ := blockMap["type"].(string); blockType == "text" {
+				if text, _ := blockMap["text"].(string); text != "" {
+					parts = append(parts, text)
+				}
+			}
+		}
+	}
+	return strings.Join(parts, "\n")
+}
+
+// setUsageAttributes sets usage-related attributes on the span.
+func setUsageAttributes(span trace.Span, usage map[string]interface{}) {
+	var inputTokens, outputTokens, totalTokens int64
+
+	if v, ok := usage["input_tokens"].(float64); ok {
+		inputTokens = int64(v)
+	}
+	if v, ok := usage["output_tokens"].(float64); ok {
+		outputTokens = int64(v)
+	}
+	if v, ok := usage["total_tokens"].(float64); ok {
+		totalTokens = int64(v)
+	}
+
+	// Handle cache tokens if present
+	var cacheCreate, cacheRead int64
+	if v, ok := usage["cache_creation_input_tokens"].(float64); ok {
+		cacheCreate = int64(v)
+	}
+	if v, ok := usage["cache_read_input_tokens"].(float64); ok {
+		cacheRead = int64(v)
+	}
+
+	// Total prompt tokens includes cache
+	totalPrompt := inputTokens + cacheCreate + cacheRead
+
+	if totalPrompt > 0 {
+		span.SetAttributes(attribute.Int64("gen_ai.usage.input_tokens", totalPrompt))
+	}
+	if outputTokens > 0 {
+		span.SetAttributes(attribute.Int64("gen_ai.usage.output_tokens", outputTokens))
+	}
+	if totalTokens > 0 {
+		span.SetAttributes(attribute.Int64("gen_ai.usage.total_tokens", totalTokens))
+	}
+
+	// Cache breakdown in metadata
+	if cacheCreate > 0 {
+		span.SetAttributes(attribute.Int64("langsmith.metadata.usage_metadata.input_token_details.cache_creation", cacheCreate))
+	}
+	if cacheRead > 0 {
+		span.SetAttributes(attribute.Int64("langsmith.metadata.usage_metadata.input_token_details.cache_read", cacheRead))
+	}
+
+	// Reasoning tokens if present (for Claude Sonnet 4.5+)
+	if v, ok := usage["reasoning_tokens"].(float64); ok {
+		span.SetAttributes(attribute.Int64("gen_ai.usage.details.reasoning_tokens", int64(v)))
+	}
+}
